addressing PR comments

+++ test/scripts/openshift-ci/teardown-e2e-setup.sh
@@ -75,11 +75,12 @@ kustomize build $PROJECT_ROOT/config/overlays/test |
   sed "s|kserve/kserve-controller:latest|${KSERVE_CONTROLLER_IMAGE}|" |
   oc delete --server-side=true -f -
 
-echo "Deleting TLS MinIO resources and generated certificates"
-kustomize build $PROJECT_ROOT/test/overlays/openshift-ci |
-  oc delete -n kserve -f -
-rm -rf $PROJECT_ROOT/test/scripts/openshift-ci/tls/certs
-
+if [ "$1" =~ "kserve_on_openshift" ]; then
+  echo "Deleting TLS MinIO resources and generated certificates"
+  kustomize build $PROJECT_ROOT/test/overlays/openshift-ci |
+    oc delete -n kserve -f -
+  rm -rf $PROJECT_ROOT/test/scripts/openshift-ci/tls/certs
+fi
 # Install DSC/DSCI for test. (sometimes there is timing issue when it is under the same kustomization so it is separated)
 oc delete -f config/overlays/test/dsci.yaml
 oc delete -f config/overlays/test/dsc.yaml
